Feature/operation result

+++ CarsInfo.Infrastructure.BusinessLogic/Services/CarsService.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using CarsInfo.Application.BusinessLogic.Contracts;
 using CarsInfo.Application.BusinessLogic.Dtos;
 using CarsInfo.Application.BusinessLogic.Enums;
-using CarsInfo.Application.BusinessLogic.Exceptions;
+using CarsInfo.Application.BusinessLogic.OperationResult;
 using CarsInfo.Application.BusinessLogic.Validators;
 using CarsInfo.Application.Persistence.Contracts;
 using CarsInfo.Application.Persistence.Filters;
@@ -19,6 +19,7 @@ namespace CarsInfo.Infrastructure.BusinessLogic.Services
     {
         private readonly ICarsRepository _carsRepository;
         private readonly IGenericRepository<CarPicture> _carsPictureRepository;
+        private readonly IUsersRepository _usersRepository;
         private readonly IGenericRepository<UserCar> _userCarRepository;
         private readonly ILogger<CarsService> _logger;
         private readonly CarServiceMapper _mapper;
@@ -27,6 +28,7 @@ namespace CarsInfo.Infrastructure.BusinessLogic.Services
         public CarsService(
             ICarsRepository carsRepository,
             IGenericRepository<CarPicture> carsPictureRepository,
+            IUsersRepository usersRepository,
             IGenericRepository<UserCar> userCarRepository,
             ILogger<CarsService> logger,
             CarServiceMapper mapper,
@@ -34,20 +36,23 @@ namespace CarsInfo.Infrastructure.BusinessLogic.Services
         {
             _carsRepository = carsRepository;
             _carsPictureRepository = carsPictureRepository;
+            _usersRepository = usersRepository;
             _userCarRepository = userCarRepository;
             _logger = logger;
             _mapper = mapper;
             _filterService = filterService;
         }
 
-        public async Task<bool> AddAsync(CarEditorDto entity)
+        public async Task<OperationResult<int>> AddAsync(CarEditorDto entity)
         {
             try
             {
-                ValidateCarEditorDto(entity);
-                var car = _mapper.MapToCar(entity);
-                var carId = await _carsRepository.AddAsync(car);
-
+                if (!entity.CarPicturesUrls?.Any() ?? true)
+                {
+                    return OperationResult<int>.FailureResult("Car should have at least one picture");
+                }
+                
+                var carId = await _carsRepository.AddAsync(_mapper.MapToCar(entity));
                 await _carsPictureRepository.AddRangeAsync(entity.CarPicturesUrls.Select(
                     carPicture => new CarPicture
                     {
@@ -55,22 +60,30 @@ namespace CarsInfo.Infrastructure.BusinessLogic.Services
                         PictureLink = carPicture
                     }).ToList());
 
-                return true;
+                return OperationResult<int>.SuccessResult(carId);
             }
             catch (Exception e)
             {
                 _logger.LogError(e, "An error occurred while creating car");
-                return false;
+                return OperationResult<int>.ExceptionResult();
             }
         }
 
-        public async Task<ToggleFavoriteStatus> ToggleFavoriteAsync(int userId, int carId)
+        public async Task<OperationResult<ToggleFavoriteStatus>> ToggleFavoriteAsync(int userId, int carId)
         {
             try
             {
                 var car = await _carsRepository.GetByIdAsync(carId);
-                ValidationHelper.ThrowIfNull(car);
+                if (car is null)
+                {
+                    return OperationResult<ToggleFavoriteStatus>.FailureResult($"Car with id={carId} does not exist");
+                }
 
+                var user = await _usersRepository.GetAsync(userId);
+                if (user is null)
+                {
+                    return OperationResult<ToggleFavoriteStatus>.FailureResult($"User with id={carId} does not exist");
+                }
                 
                 var userCar = await _userCarRepository.GetAsync(new List<FiltrationField>
                 {
@@ -80,19 +93,21 @@ namespace CarsInfo.Infrastructure.BusinessLogic.Services
 
                 if (userCar is null || userCar.IsDeleted)
                 {
-                    return await AddToFavoriteAsync(userCar, carId, userId);
+                    await AddToFavoriteAsync(userCar, carId, userId);
+                    return OperationResult<ToggleFavoriteStatus>.SuccessResult(ToggleFavoriteStatus.AddedToFavorite);
                 }
 
-                return await DeleteFromFavoriteAsync(userCar.Id);
+                await DeleteFromFavoriteAsync(userCar.Id);
+                return OperationResult<ToggleFavoriteStatus>.SuccessResult(ToggleFavoriteStatus.DeleteFromFavorite);
             }
             catch (Exception e)
             {
                 _logger.LogError(e, "An error occurred while adding car with to favorite");
-                return ToggleFavoriteStatus.Error;
+                return OperationResult<ToggleFavoriteStatus>.ExceptionResult();
             }
         }
 
-        private async Task<ToggleFavoriteStatus> AddToFavoriteAsync(UserCar userCar, int carId, int userId)
+        private async Task AddToFavoriteAsync(UserCar userCar, int carId, int userId)
         {
             if (userCar is null)
             {
@@ -101,48 +116,53 @@ namespace CarsInfo.Infrastructure.BusinessLogic.Services
                     CarId = carId,
                     UserId = userId
                 });
-                return ToggleFavoriteStatus.AddedToFavorite;
+                return;
             }
 
             userCar.IsDeleted = false;
             await _userCarRepository.UpdateAsync(userCar);
-            return ToggleFavoriteStatus.AddedToFavorite;
         }
 
-        private async Task<ToggleFavoriteStatus> DeleteFromFavoriteAsync(int userCarId)
+        private async Task DeleteFromFavoriteAsync(int userCarId)
         {
             await _userCarRepository.DeleteAsync(userCarId);
-            return ToggleFavoriteStatus.DeleteFromFavorite;
         }
 
-        public async Task DeleteByIdAsync(int id)
+        public async Task<OperationResult> DeleteByIdAsync(int id)
         {
             try
             {
+                var car = await _carsRepository.GetByIdAsync(id);
+                if (car is null)
+                {
+                    return OperationResult.FailureResult($"Car with id={id} does not exist");
+                }
+                
                 await _carsRepository.DeleteAsync(id);
+                return OperationResult.SuccessResult();
             }
             catch (Exception e)
             {
                 _logger.LogError(e, $"An error occurred while deleting car with id={id}");
+                return OperationResult.ExceptionResult();
             }
         }
 
-        public async Task<IEnumerable<CarDto>> GetAllAsync()
+        public async Task<OperationResult<IEnumerable<CarDto>>> GetAllAsync()
         {
             try
             {
                 var cars = await _carsRepository.GetAsync();
-                var carsDtos = _mapper.MapToCarsDtos(cars);
-                return carsDtos;
+                return OperationResult<IEnumerable<CarDto>>.SuccessResult(_mapper.MapToCarsDtos(cars));
             }
             catch (Exception e)
             {
                 _logger.LogError(e, "An error occurred while fetching all cars");
-                return new List<CarDto>();
+                return OperationResult<IEnumerable<CarDto>>.ExceptionResult();
             }
         }
 
-        public async Task<IEnumerable<CarDto>> GetAllAsync(FilterDto filterDto)
+        public async Task<OperationResult<IEnumerable<CarDto>>> GetAllAsync(FilterDto filterDto)
         {
             try
             {
@@ -153,116 +173,117 @@ namespace CarsInfo.Infrastructure.BusinessLogic.Services
 
                 var filter = _filterService.ConfigureCarFilter(filterDto);
                 var cars = await _carsRepository.GetAsync(filter);
-                var carsDtos = _mapper.MapToCarsDtos(cars);
-
-                return carsDtos;
+                return OperationResult<IEnumerable<CarDto>>.SuccessResult(_mapper.MapToCarsDtos(cars));
             }
             catch (Exception e)
             {
                 _logger.LogError(e, "An error occurred while fetching all cars");
-                return new List<CarDto>();
+                return OperationResult<IEnumerable<CarDto>>.ExceptionResult();
             }
         }
 
-        public async Task<IEnumerable<CarDto>> GetUserFavoriteCarsAsync(int userId, FilterDto filter)
+        public async Task<OperationResult<IEnumerable<CarDto>>> GetUserFavoriteCarsAsync(int userId, FilterDto filter)
         {
             try
             {
-                ValidationHelper.ThrowIfNull(filter);
+                var user = await _usersRepository.GetAsync(userId);
+                if (user is null)
+                {
+                    return OperationResult<IEnumerable<CarDto>>.FailureResult($"User with id={userId} does not exist");
+                }
 
                 var filterModel = _filterService.ConfigureCarFilter(filter);
                 filterModel.Filters.Add(new FiltrationField("UserCar.IsDeleted", 0));
                 var cars = await _carsRepository.GetUserFavoriteCarsAsync(userId, filterModel);
-                var carsDtos = _mapper.MapToCarsDtos(cars).ToList();
 
-                return carsDtos;
+                return OperationResult<IEnumerable<CarDto>>.SuccessResult(_mapper.MapToCarsDtos(cars));
             }
             catch (Exception e)
             {
                 _logger.LogError(e, "An error occurred while fetching all cars");
-                return new List<CarDto>();
+                return OperationResult<IEnumerable<CarDto>>.ExceptionResult();
             }
         }
 
-        public async Task<IEnumerable<int>> GetUserFavoriteCarsIdsAsync(int userId)
+        public async Task<OperationResult<IEnumerable<int>>> GetUserFavoriteCarsIdsAsync(int userId)
         {
             try
             {
-                var filterModel = new FilterModel();
-                filterModel.Filters.Add(new FiltrationField("UserCar.IsDeleted", 0));
+                var user = await _usersRepository.GetAsync(userId);
+                if (user is null)
+                {
+                    return OperationResult<IEnumerable<int>>.FailureResult($"User with id={userId} does not exist");
+                }
+
+                var filterModel = new FilterModel(new FiltrationField("UserCar.IsDeleted", 0));
                 var carsIds = await _carsRepository.GetUserFavoriteCarsIdsAsync(userId, filterModel);
-                return carsIds;
+
+                return OperationResult<IEnumerable<int>>.SuccessResult(carsIds);
             }
             catch (Exception e)
             {
                 _logger.LogError(e, "An error occurred while fetching favorite cars ids");
-                return new List<int>();
+                return OperationResult<IEnumerable<int>>.ExceptionResult();
             }
         }
-
-        public async Task<CarDto> GetByIdAsync(int id)
+        
+        public async Task<OperationResult<CarDto>> GetByIdAsync(int id)
         {
             try
             {
                 var car = await _carsRepository.GetByIdAsync(id);
-                var carDto = _mapper.MapToCarDto(car);
-
-                return carDto;
+                return OperationResult<CarDto>.SuccessResult(_mapper.MapToCarDto(car));
             }
             catch (Exception e)
             {
                 _logger.LogError(e, $"An error occurred while fetching car with id={id}");
-                return null;
+                return OperationResult<CarDto>.ExceptionResult();
             }
         }
 
-        public async Task<CarEditorDto> GetCarEditorDtoByIdAsync(int id)
+        public async Task<OperationResult<CarEditorDto>> GetCarEditorDtoByIdAsync(int id)
         {
             try
             {
                 var car = await _carsRepository.GetByIdAsync(id);
-                var carDto = _mapper.MapToCarEditorDto(car);
-
-                return carDto;
+                return OperationResult<CarEditorDto>.SuccessResult(_mapper.MapToCarEditorDto(car));
             }
             catch (Exception e)
             {
                 _logger.LogError(e, $"An error occurred while fetching car with id={id}");
-                return null;
+                return OperationResult<CarEditorDto>.ExceptionResult();
             }
         }
 
-        public async Task UpdateAsync(CarEditorDto entity)
+        public async Task<OperationResult> UpdateAsync(CarEditorDto entity)
         {
             try
             {
-                ValidateCarEditorDto(entity);
-                var car = _mapper.MapToCar(entity);
-                var oldCar = await _carsRepository.GetByIdAsync(entity.Id);
-
-                await _carsRepository.UpdateAsync(car);
-                await _carsPictureRepository.DeleteRangeAsync(oldCar.CarPictures.Select(cp => cp.Id));
+                if (!entity.CarPicturesUrls?.Any() ?? true)
+                {
+                    return OperationResult.FailureResult("Car should have at least one picture");
+                }
+                
+                var car = await _carsRepository.GetByIdAsync(entity.Id);
+                if (car is null)
+                {
+                    return OperationResult.FailureResult($"Car with id={entity.Id} does not exist");
+                }
+                
+                await _carsRepository.UpdateAsync(_mapper.MapToCar(entity));
+                await _carsPictureRepository.DeleteRangeAsync(car.CarPictures.Select(cp => cp.Id));
                 await _carsPictureRepository.AddRangeAsync(entity.CarPicturesUrls.Select(
                     carPicture => new CarPicture
                     {
-                        CarId = car.Id,
+                        CarId = entity.Id,
                         PictureLink = carPicture
                     }).ToList());
+                return OperationResult.SuccessResult();
             }
             catch (Exception e)
             {
                 _logger.LogError(e, $"An error occurred while updating car with id={entity.Id}");
-            }
-        }
-
-        private static void ValidateCarEditorDto(CarEditorDto car)
-        {
-            ValidationHelper.ThrowIfNull(car);
-            ValidationHelper.ThrowIfStringNullOrWhiteSpace(car.Model);
-
-            if (!car.CarPicturesUrls?.Any() ?? true)
-            {
-                throw new BllException("Car should have at least one picture");
+                return OperationResult.ExceptionResult();
             }
         }
     }
